tidy up a bit
